Add JUnit tests for new FindCommand
