+++ src/main/java/seedu/address/logic/commands/FindCommand.java
@@ -1,35 +1,28 @@
 package seedu.address.logic.commands;
 
-import static java.util.Objects.requireNonNull;
-
 import seedu.address.commons.core.Messages;
 import seedu.address.model.Model;
-import seedu.address.model.person.NameContainsKeywordsPredicate;
+import seedu.address.model.person.PersonContainsKeywordPredicate;
 
-/**
- * Finds and lists all persons in address book whose name contains any of the argument keywords.
- * Keyword matching is case insensitive.
- */
-public class FindCommand extends Command {
+import static java.util.Objects.requireNonNull;
 
-    public static final String COMMAND_WORD = "find";
+public class SpecialFindCommand extends Command {
+    public static final String COMMAND_WORD = "sfind";
 
-    public static final String MESSAGE_USAGE = COMMAND_WORD + ": Finds all persons whose names contain any of "
+    public static final String MESSAGE_USAGE = COMMAND_WORD + ": Finds all persons whose contact details matches any of "
             + "the specified keywords (case-insensitive) and displays them as a list with index numbers.\n"
             + "Parameters: KEYWORD [MORE_KEYWORDS]...\n"
-            + "Example: " + COMMAND_WORD + " alice bob charlie";
+            + "Example: " + COMMAND_WORD + " n/ alice bob charlie" + " e/ alice@gmail.com";
 
-    private final NameContainsKeywordsPredicate predicate;
+    private final PersonContainsKeywordPredicate predicate;
 
-    public FindCommand(NameContainsKeywordsPredicate predicate) {
+    public SpecialFindCommand(PersonContainsKeywordPredicate predicate) {
         this.predicate = predicate;
     }
 
     @Override
     public CommandResult execute(Model model) {
         requireNonNull(model);
-        // Update model with the filter person list.
-        // Note: Everytime a find command is called, a new filtered list is generated in ModelManager class.
         model.updateFilteredPersonList(predicate);
         return new CommandResult(
                 String.format(Messages.MESSAGE_PERSONS_LISTED_OVERVIEW, model.getFilteredPersonList().size()));
@@ -38,7 +31,7 @@ public class FindCommand extends Command {
     @Override
     public boolean equals(Object other) {
         return other == this // short circuit if same object
-                || (other instanceof FindCommand // instanceof handles nulls
-                && predicate.equals(((FindCommand) other).predicate)); // state check
+                || (other instanceof SpecialFindCommand // instanceof handles nulls
+                && predicate.equals(((SpecialFindCommand) other).predicate)); // state check
     }
 }
