--- src/main/java/seedu/address/logic/parser/SpecialFindCommandParser.java
@@ -1,35 +1,0 @@
-package seedu.address.logic.parser;
-
-import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
-import static seedu.address.logic.parser.CliSyntax.*;
-
-import seedu.address.logic.commands.SpecialFindCommand;
-import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.person.PersonContainsKeywordPredicate;
-
-
-/**
- * Parses input arguments and creates a new FindCommand object
- */
-public class SpecialFindCommandParser implements Parser<SpecialFindCommand> {
-
-    /**
-     * Parses the given {@code String} of arguments in the context of the FindCommand
-     * and returns a FindCommand object for execution.
-     * @throws ParseException if the user input does not conform the expected format
-     */
-    public SpecialFindCommand parse(String args) throws ParseException {
-        String trimmedArgs = args.trim();
-        if (trimmedArgs.isEmpty()) {
-            throw new ParseException(
-                    String.format(MESSAGE_INVALID_COMMAND_FORMAT, SpecialFindCommand.MESSAGE_USAGE));
-        }
-
-        ArgumentMultimap argMultimap =
-                ArgumentTokenizer.tokenize(args, PREFIX_NAME, PREFIX_PHONE, PREFIX_EMAIL, PREFIX_ADDRESS, PREFIX_TAG);
-
-        PersonContainsKeywordPredicate predicate =
-                new PersonContainsKeywordPredicate((argMultimap));
-        return new SpecialFindCommand(predicate);
-    }
-}
